requirements.txt
pytest
pytest-html  # Untuk laporan HTML (opsional)
pytest-xdist  # Untuk parallel testing
requests
jsonschema
python-dotenv
faker
allure-pytest  # Tambahkan Allure untuk reporting
